refactor(gateway): renaming rpc to gateway, moving to /api/
