+++ api/gateway/server_test.go
@@ -1,4 +1,4 @@
-package rpc
+package gateway
 
 import (
 	"context"
