+++ nodebuilder/rpc/flags.go
@@ -4,12 +4,12 @@ import (
 	"github.com/spf13/cobra"
 	flag "github.com/spf13/pflag"
 
-	"github.com/celestiaorg/celestia-node/service/rpc"
+	"github.com/celestiaorg/celestia-node/api/gateway"
 )
 
 var (
-	addrFlag = "rpc.addr"
-	portFlag = "rpc.port"
+	addrFlag = "gateway.addr"
+	portFlag = "gateway.port"
 )
 
 // Flags gives a set of hardcoded node/rpc package flags.
@@ -31,7 +31,7 @@ func Flags() *flag.FlagSet {
 }
 
 // ParseFlags parses RPC flags from the given cmd and saves them to the passed config.
-func ParseFlags(cmd *cobra.Command, cfg *rpc.Config) {
+func ParseFlags(cmd *cobra.Command, cfg *gateway.Config) {
 	addr := cmd.Flag(addrFlag).Value.String()
 	if addr != "" {
 		cfg.Address = addr
